+++ src/pages/GamePage/GameMatchingPage/components/invite-friend-dialog/styles.css.ts
